Revert "renamed to policy evaluator"
This reverts commit 3d40489e3a79eed5e2a53dbb26c23e91bc8a08ae.
