+++ python/ray/rllib/evaluation/__init__.py
@@ -1,5 +1,5 @@
-from ray.rllib.evaluation.policy_evaluator import (
-    PolicyEvaluator, PolicyEvaluator)
+from ray.rllib.evaluation.common_policy_evaluator import (
+    CommonPolicyEvaluator, PolicyEvaluator)
 from ray.rllib.evaluation.interface import EvaluatorInterface
 from ray.rllib.evaluation.policy_graph import PolicyGraph
 from ray.rllib.evaluation.tf_policy_graph import TFPolicyGraph
@@ -11,7 +11,8 @@ from ray.rllib.evaluation.postprocessing import compute_advantages
 from ray.rllib.evaluation.metrics import collect_metrics
 
 __all__ = [
-    "EvaluatorInterface", "PolicyEvaluator", "PolicyGraph", "TFPolicyGraph",
+    "EvaluatorInterface", "CommonPolicyEvaluator", "PolicyEvaluator",
+    "PolicyGraph", "TFPolicyGraph",
     "TorchPolicyGraph", "SampleBatch", "MultiAgentBatch", "SampleBatchBuilder",
     "MultiAgentSampleBatchBuilder", "SyncSampler", "AsyncSampler",
     "compute_advantages", "collect_metrics"
