+++ python/ray/rllib/evaluation/metrics.py
@@ -10,7 +10,7 @@ from ray.tune.result import TrainingResult
 
 
 def collect_metrics(local_evaluator, remote_evaluators=[]):
-    """Gathers episode metrics from PolicyEvaluator instances."""
+    """Gathers episode metrics from CommonPolicyEvaluator instances."""
 
     episode_rewards = []
     episode_lengths = []
