+++ python/ray/rllib/test/test_common_policy_evaluator.py
@@ -8,7 +8,7 @@ import unittest
 
 import ray
 from ray.rllib.agents.pg import PGAgent
-from ray.rllib.evaluation.policy_evaluator import PolicyEvaluator
+from ray.rllib.evaluation.common_policy_evaluator import CommonPolicyEvaluator
 from ray.rllib.evaluation.metrics import collect_metrics
 from ray.rllib.evaluation.policy_graph import PolicyGraph
 from ray.rllib.evaluation.postprocessing import compute_advantages
@@ -89,9 +89,9 @@ class MockVectorEnv(VectorEnv):
         return obs_batch, rew_batch, done_batch, info_batch
 
 
-class TestPolicyEvaluator(unittest.TestCase):
+class TestCommonPolicyEvaluator(unittest.TestCase):
     def testBasic(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: gym.make("CartPole-v0"),
             policy_graph=MockPolicyGraph)
         batch = ev.sample()
@@ -110,10 +110,10 @@ class TestPolicyEvaluator(unittest.TestCase):
         self.assertEqual(results2, [(0, 5), (1, 5), (2, 5)])
 
     def testMetrics(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: MockEnv(episode_length=10),
             policy_graph=MockPolicyGraph, batch_mode="complete_episodes")
-        remote_ev = PolicyEvaluator.as_remote().remote(
+        remote_ev = CommonPolicyEvaluator.as_remote().remote(
             env_creator=lambda _: MockEnv(episode_length=10),
             policy_graph=MockPolicyGraph, batch_mode="complete_episodes")
         ev.sample()
@@ -123,7 +123,7 @@ class TestPolicyEvaluator(unittest.TestCase):
         self.assertEqual(result.episode_reward_mean, 10)
 
     def testAsync(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: gym.make("CartPole-v0"),
             sample_async=True,
             policy_graph=MockPolicyGraph)
@@ -133,7 +133,7 @@ class TestPolicyEvaluator(unittest.TestCase):
         self.assertGreater(batch["advantages"][0], 1)
 
     def testAutoConcat(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: MockEnv(episode_length=40),
             policy_graph=MockPolicyGraph,
             sample_async=True,
@@ -145,7 +145,7 @@ class TestPolicyEvaluator(unittest.TestCase):
         self.assertEqual(batch.count, 40)  # auto-concat up to 5 episodes
 
     def testAutoVectorization(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: MockEnv(episode_length=20),
             policy_graph=MockPolicyGraph,
             batch_mode="truncate_episodes",
@@ -164,14 +164,14 @@ class TestPolicyEvaluator(unittest.TestCase):
     def testBatchDivisibilityCheck(self):
         self.assertRaises(
             ValueError,
-            lambda: PolicyEvaluator(
+            lambda: CommonPolicyEvaluator(
                 env_creator=lambda _: MockEnv(episode_length=8),
                 policy_graph=MockPolicyGraph,
                 batch_mode="truncate_episodes",
                 batch_steps=15, num_envs=4))
 
     def testBatchesSmallerWhenVectorized(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: MockEnv(episode_length=8),
             policy_graph=MockPolicyGraph,
             batch_mode="truncate_episodes",
@@ -185,7 +185,7 @@ class TestPolicyEvaluator(unittest.TestCase):
         self.assertEqual(result.episodes_total, 4)
 
     def testVectorEnvSupport(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: MockVectorEnv(
                 episode_length=20, num_envs=8),
             policy_graph=MockPolicyGraph,
@@ -203,7 +203,7 @@ class TestPolicyEvaluator(unittest.TestCase):
         self.assertEqual(result.episodes_total, 8)
 
     def testTruncateEpisodes(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: MockEnv(10),
             policy_graph=MockPolicyGraph,
             batch_steps=15,
@@ -212,7 +212,7 @@ class TestPolicyEvaluator(unittest.TestCase):
         self.assertEqual(batch.count, 15)
 
     def testCompleteEpisodes(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: MockEnv(10),
             policy_graph=MockPolicyGraph,
             batch_steps=5,
@@ -221,7 +221,7 @@ class TestPolicyEvaluator(unittest.TestCase):
         self.assertEqual(batch.count, 10)
 
     def testCompleteEpisodesPacking(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: MockEnv(10),
             policy_graph=MockPolicyGraph,
             batch_steps=15,
@@ -233,7 +233,7 @@ class TestPolicyEvaluator(unittest.TestCase):
             [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9])
 
     def testFilterSync(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: gym.make("CartPole-v0"),
             policy_graph=MockPolicyGraph,
             sample_async=True,
@@ -246,7 +246,7 @@ class TestPolicyEvaluator(unittest.TestCase):
         self.assertNotEqual(obs_f.buffer.n, 0)
 
     def testGetFilters(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: gym.make("CartPole-v0"),
             policy_graph=MockPolicyGraph,
             sample_async=True,
@@ -261,7 +261,7 @@ class TestPolicyEvaluator(unittest.TestCase):
         self.assertGreaterEqual(obs_f2.buffer.n, obs_f.buffer.n)
 
     def testSyncFilter(self):
-        ev = PolicyEvaluator(
+        ev = CommonPolicyEvaluator(
             env_creator=lambda _: gym.make("CartPole-v0"),
             policy_graph=MockPolicyGraph,
             sample_async=True,
